Interface for userDao added
