+++ bin/.gitignore
@@ -1,0 +1,39 @@
+
+
+# Compiled class file
+*.class
+
+# Log file
+*.log
+
+# BlueJ files
+*.ctxt
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
+#Removing target folder
+target/**
+
+.metadata/**
+
+.factorypath
+
+.classpath
+
+.vscode/**
+
+
+.settings/**
