+++ bin/README.md
@@ -1,0 +1,6 @@
+# ServeD2D
+
+## Using an online SQL Based Database
+  * In the development process of the project, we will be using a database from https://www.freesqldatabase.com/.
+  * Currently a DB server is running at http://sql6.freesqldatabase.com:3306/.
+  * Login credentials will be provided personally.
