+++ bin/requests.http
@@ -1,0 +1,210 @@
+POST /api/login
+Host: localhost:8080
+Content-Type: application/json
+
+{
+    "id" : 1,
+    "name" : "Zonnya Lorraine",
+    "isProvider" : "0",
+    "walletBalance" : "5.40",
+    "email": "zlorraine0@devhub.com",
+    "password": "MB86fpAFUS0"
+}
+### 
+
+POST /api/register
+Host: localhost:8080
+Content-Type: application/json
+
+{
+    "id" : 21,
+    "name" : "Subham Srivastava",
+    "isProvider" : "0",
+    "walletBalance" : "12.23",
+    "email": "subhamxx@gmail.com",
+    "password": "1234"
+}
+
+### ORDERS ----------------------------------------------------------------------------------------------
+
+GET /api/orders
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+GET /api/orders/1
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+GET /api/orders/21
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+GET /api/orders?status=2
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+POST /api/orders
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+Content-Type: application/json
+
+{
+    "orderId" : 21,
+    "userId" : 1,
+    "serviceId" : 2,
+	"adressId" : 1,
+	"timestamp" : "2013-01-20",
+	"orderStatus" : 3,
+	"amount" : 23.43
+}
+
+###
+
+POST /api/orders?order_id=21
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+Content-Type: application/json
+
+{
+    "orderId" : 21,
+    "userId" : 1,
+    "serviceId" : 2,
+	"adressId" : 1,
+	"timestamp" : "2013-01-20",
+	"orderStatus" : 2,
+	"amount" : 23.43
+}
+
+### SERVICES ----------------------------------------------------------------------------------------------
+
+GET /api/services
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+GET /api/services/categories
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+GET /api/services?category=consectetuer
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+GET /api/services/2
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+POST /api/services
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+Content-Type: application/json
+
+{
+    "serviceId": 21,
+    "providerId": 2,
+    "addressId": 2,
+    "isApproved": true,
+    "category": "mattis",
+    "description": "Morbi porttitor lorem id ligula. Suspendisse ornare consequat lectus. In est risus, auctor sed, tristique in, tempus sit amet, sem.",
+    "imageUrl": "http://dummyimage.com/122x100.png/ff4444/ffffff",
+    "serviceRadius": 10.0,
+    "price": 25.1,
+    "ratingCount": 1000000,
+    "completedOrders": 2222
+}
+
+###
+
+POST /api/services/edit
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+Content-Type: application/json
+
+{
+    "serviceId": 21,
+    "providerId": 2,
+    "addressId": 2,
+    "isApproved": true,
+    "category": "consectetuer",
+    "description": "Morbi porttitor lorem id ligula. Suspendisse ornare consequat lectus. In est risus, auctor sed, tristique in, tempus sit amet, sem.",
+    "imageUrl": "http://dummyimage.com/122x100.png/ff4444/ffffff",
+    "serviceRadius": 10.0,
+    "price": 25.1,
+    "ratingCount": 1000000,
+    "completedOrders": 2222
+}
+
+### PAYMENT ----------------------------------------------------------------------------------------------
+
+GET /api/wallet/balance
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+
+###
+
+POST /api/wallet/balance/add
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+Amount : 20
+
+###
+
+POST /api/wallet/balance/withdraw
+Host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTU0NjgxLCJpYXQiOjE2MjA4OTQ2ODF9.7GQHOAmXvqlTemX7KOaifY6e7xpledhWDWtXzn8eR1g
+Amount : 10
+
+###
+
+POST /api/wallet/balance/payment
+host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTc3MDQ4LCJpYXQiOjE2MjA5MTcwNDh9.U1fUoax8VqeJWLUFfGUjEBxVayHsmBJrsmF5knKVqxw
+Content-Type: application/json
+
+{
+    "orderId" : 1,
+    "userId" : 1,
+    "serviceId" : 2,
+	"adressId" : 1,
+	"timestamp" : "2013-01-20",
+	"orderStatus" : 3,
+	"amount" : 30.90
+}
+
+
+### ADDRESS ------------------------------------------------------------------------
+
+GET /api/address
+host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTc3NzM3LCJpYXQiOjE2MjA5MTc3Mzd9.ujyWL2wFwXBI4dAHj2Izh4E0rae5jlfIJG3zDmEwUx0
+
+###
+
+POST /address/edit
+host: localhost:8080
+Authorization: JWT eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJab25ueWEgTG9ycmFpbmUiLCJpZCI6MSwiZXhwIjoxNjIwOTc5MjEwLCJpYXQiOjE2MjA5MTkyMTB9.hqD91BmW0_6hjxtFzjYFxyHDoqNITz_SYQsEy1Sj8eI
+Content-Type: application/json
+
+{
+  "addressid": 1,
+  "userId": 1,
+  "lat": 50.7557791,
+  "longi": 38.5709037
+}
+
+
