+++ bin/src/main/resources/jdbc-data.properties
@@ -1,0 +1,5 @@
+  
+jdbc.driverClassName=com.mysql.cj.jdbc.Driver
+jdbc.username=sql6409360
+jdbc.password=GwFlfLvu4P
+jdbc.url=jdbc:mysql://sql6.freesqldatabase.com:3306/sql6409360
