+++ bin/src/main/resources/secret-key.properties
@@ -1,0 +1,1 @@
+key.secretKey=ZL8KN.k?kpyvCnKQ.f$eDh%e^NvBX.6yAVgNVB2aP:@b/<kz
